refactor: remove subcomponents folder from pagination folder
